# api/openapi-lock.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "org-quicko-cliq",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://{{base_url}}"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "paths": {
...
    "/programs": {
      "post": {
        "tags": [
          "Program"
        ],
        "summary": "Create program",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "program"
                  },
                  "program_id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "currency": {
                    "type": "string"
                  },
                  "website": {
                    "type": "string"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "public",
                      "private"
                    ]
                  }
                },
                "required": [
                  "name",
                  "currency",
                  "website",
                  "visibility"
                ]
              },
              "examples": {
                "success": {
                  "value": {
                    "name": "Influencer program",
                    "currency": "USD",
                    "visibility": "public",
                    "website": "www.example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully created program.\",\r\n    \"data\": {\r\n        \"program_id\": \"ewvkj2n73c8-=3=-h91bjb*^#@678$2\"\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {
                  "201-Example-1": {
                    "value": {
                      "name": "Beginners",
                      "currency": "USD",
                      "visibility": "PUBLIC",
                      "website": "www.example.com"
                    },
                    "summary": "Example 1 for status 201"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Program"
        ],
        "summary": "Get all programs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully got all programs.\",\r\n    \"data\": {\r\n        \"programs\": [\r\n            {\r\n                \"name\": \"StockBroker program\",\r\n                \"currency\": \"INR\",\r\n                \"visibility\": \"public\",\r\n                \"website\": \"www.example.com\"\r\n            },\r\n            {\r\n                \"name\": \"Bankers program\",\r\n                \"currency\": \"EUR\",\r\n                \"visibility\": \"public\",\r\n                \"website\": \"www.mockXplain.com\"\r\n            }\r\n        ]\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/programs/{program_id}": {
      "get": {
        "tags": [
          "Program"
        ],
        "summary": "Get program",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          }
        ],
        "responses": {
          "undefined": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully got program data.\",\r\n    \"data\": {\r\n        \"program\": {\r\n            \"program_id\": \"ewvkj2n73c8-=3=-h91bjb*^#@678$2\",\r\n            \"name\": \"Influencer program\",\r\n            \"website\": \"www.example.com\",\r\n            \"visibility\": \"public\",\r\n            \"currency\": \"USD\"\r\n        }\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Program"
        ],
        "summary": "Update program",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "program"
                  },
                  "program_id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "currency": {
                    "type": "string"
                  },
                  "website": {
                    "type": "string"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "public",
                      "private"
                    ]
                  }
                },
                "required": [
                  "name",
                  "currency",
                  "website",
                  "visibility"
                ]
              },
              "examples": {
                "success": {
                  "value": {
                    "name": "StockBroker program",
                    "currency": "INR",
                    "visibility": "public",
                    "website": "www.example.com"
                  }
                },
                "failure- unauthorized": {
                  "value": {
                    "id": "",
                    "name": "",
                    "currency": "",
                    "visibility": "",
                    "website": ""
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully updated program.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"error\": \"You are unauthorized to perform this action.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Program"
        ],
        "summary": "Delete program",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully deleted program.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"error\": \"You are unauthorized to perform this action.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/programs/{program_id}/invite": {
      "post": {
        "tags": [
          "Program"
        ],
        "summary": "Invite user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "user"
                  },
                  "user_id": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "member"
                    ]
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "examples": {
                "success": {
                  "value": {
                    "email": "john@mail.com",
                    "password": "knw97890un",
                    "first_name": "John",
                    "last_name": "Bennington",
                    "role": "MEMBER"
                  }
                },
                "failure- user already part of program": {
                  "value": {
                    "email": "john@mail.com",
                    "password": "knw97890un",
                    "first_name": "John",
                    "last_name": "Bennington",
                    "role": "MEMBER"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully invited user to program\",\r\n    \"data\": {\r\n        \"user_id\": \"r3y10jij&78=-=-%^&%%\"\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"error\": \"Invalid request. User is already part of the program.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/programs/{program_id}/users": {
      "get": {
        "tags": [
          "Program"
        ],
        "summary": "Get all users",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully got all program users.\",\r\n    \"data\": {\r\n        \"users\": [\r\n            {\r\n                \"user_id\": \"123\",\r\n                \"email\": \"john@mail.com\",\r\n                \"first_name\": \"John\",\r\n                \"last_name\": \"Bennington\"\r\n            },\r\n            {\r\n                \"user_id\": \"456\",\r\n                \"email\": \"jane@mail.com\",\r\n                \"first_name\": \"Jane\",\r\n                \"last_name\": \"Doe\"\r\n            }\r\n        ]\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/programs/{program_id}/users/{user_id}": {
      "patch": {
        "tags": [
          "Program"
        ],
        "summary": "Update role",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "role": ""
                }
              },
              "examples": {
                "success": {
                  "value": {
                    "role": "ADMIN"
                  }
                },
                "failure- unauthorized": {
                  "value": {
                    "role": ""
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          },
          {
            "name": "user_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{user_id}}"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully changed user role.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"error\": \"You are unauthorized to perform this action.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Program"
        ],
        "summary": "Remove user",
        "description": "Mark user status as inactive",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          },
          {
            "name": "user_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{user_id}}"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully removed user from program.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"error\": \"You are unauthorized to perform this action.\"\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
...
    "/programs/{program_id}/purchases": {
      "get": {
        "tags": [
          "Program"
        ],
        "summary": "Get purchases in workspace",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully got all purchases of program.\",\r\n    \"data\": {\r\n        \"purchases\": [\r\n            {\r\n                \"purchase_id\": \"dunewcdf8927yr8ikh\",\r\n                \"email\": \"beethoven@mail.com\",\r\n                \"first_name\": \"Ludwig\",\r\n                \"last_name\": \"Van Beethoven\",\r\n                \"phone\": \"1234567899\"\r\n            },\r\n            {\r\n                \"purchase_id\": \"dunewcdf8927yr8ikh\",\r\n                \"email\": \"leonardo@mail.com\",\r\n                \"first_name\": \"Leonardo\",\r\n                \"last_name\": \"Da Vinci\",\r\n                \"phone\": \"1234567890\"\r\n            }\r\n        ]\r\n    }\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/programs/{program_id}/commissions": {
      "get": {
        "tags": [
          "Program"
        ],
        "summary": "Get all commissions",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{program_id}}"
          }
        ],
        "responses": {
          "undefined": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "{\r\n    \"message\": \"Successfully got all commissions for program.\",\r\n    \"data\": [\r\n        \"commissions\": [\r\n            {\r\n                \"commission_id\": \"123\",\r\n                \"conversion_id\": \"233v4g23f\",\r\n                \"conversion_type\": \"CONTACT\",\r\n                \"amount\": 50\r\n            },\r\n            {\r\n                \"commission_id\": \"457\",\r\n                \"conversion_id\": \"23983hc9cn\",\r\n                \"conversion_type\": \"PURCHASE\",\r\n                \"amount\": 80\r\n            }\r\n        ]\r\n    ]\r\n}"
              },
              "application/json": {
                "schema": {},
                "examples": {}
              }
            }
          }
        }
      }
    },
...
  }
}